Fix double to float (issue 33)

+++ libraries/FastMap/FastMap.cpp
@@ -1,13 +1,14 @@
 //
 //    FILE: FastMap.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.7
+// VERSION: 0.1.8
 // PURPOSE: class with fast map function - library for Arduino
 //     URL: http://forum.arduino.cc/index.php?topic=276194
 //
 // HISTORY:
+// 0.1.8  2017-07-27 revert double to float (issue 33)
 // 0.1.7  2017-04-28 cleaned up, get examples working again
-// 0.1.06 2015-03-08 replaced float by double (support ARM)
+// 0.1.06 2015-03-08 replaced float by float (support ARM)
 // 0.1.05 2014-11-02 stripped of bit mask experimental code
 // 0.1.04 add back() - the inverse map
 //        tested with bit mask for constrain code (Perfomance was killed)
@@ -28,7 +29,7 @@ FastMap::FastMap()
     init(0, 1, 0, 1);
 }
 
-void FastMap::init(double in_min, double in_max, double out_min, double out_max)
+void FastMap::init(float in_min, float in_max, float out_min, float out_max)
 {
     _in_min = in_min;
     _in_max = in_max;
@@ -42,24 +43,23 @@ void FastMap::init(double in_min, double in_max, double out_min, double out_max)
     _backbase = in_min - out_min * _backfactor;
 }
 
-double FastMap::constrainedMap(double value)
+float FastMap::constrainedMap(float value)
 {
     if (value <= _in_min) return _out_min;
     if (value >= _in_max) return _out_max;
     return this->map(value);
 }
 
-double FastMap::lowerConstrainedMap(double value)
+float FastMap::lowerConstrainedMap(float value)
 {
     if (value <= _in_min) return _out_min;
     return this->map(value);
 }
 
-double FastMap::upperConstrainedMap(double value)
+float FastMap::upperConstrainedMap(float value)
 {
     if (value >= _in_max) return _out_max;
     return this->map(value);
 }
-//
+
 // END OF FILE
-//
